Add helper functions in UnifiedCacheManager
